:sparkles: Add new MCBNs and refactor
